requirements.txt
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1
scikit-learn==1.3.0
PyYAML==6.0
pytest==7.4.0
matplotlib==3.10.8
seaborn==0.13.2
torch==2.5.1
Pillow==12.2.0
pypdfium2==5.7.0
markdown-it-py==4.0.0
